+++ src/view/plandetail/NearbyPlaceList.tsx
@@ -1,11 +1,9 @@
-import { Box, Center, HStack, Icon, Text } from "@chakra-ui/react";
-import { ReactNode, useRef } from "react";
-import { MdArrowBackIos, MdArrowForwardIos } from "react-icons/md";
+import { Box, Text } from "@chakra-ui/react";
 import { Place } from "src/domain/models/Place";
 import { createArrayWithSize } from "src/domain/util/array";
+import { HorizontaScrollablelList } from "src/view/common/HorizontaScrollablelList";
 import { ImageWithSkeleton } from "src/view/common/ImageWithSkeleton";
 import { Size } from "src/view/constants/size";
-import { isPC } from "src/view/constants/userAgent";
 
 type Props = {
     places: Place[] | null;
@@ -15,17 +13,17 @@ type Props = {
 export const NearbyPlaceList = ({ places, onSelectPlace }: Props) => {
     if (!places) {
         return (
-            <Container>
+            <HorizontaScrollablelList>
                 {createArrayWithSize(5).map((_, index) => (
                     <NearbyPlaceCardSkeleton key={index} />
                 ))}
-            </Container>
+            </HorizontaScrollablelList>
         );
     }
 
     // TODO: 要素が一つもないときの対応
     return (
-        <Container>
+        <HorizontaScrollablelList>
             {places
                 .filter((p) => p.images.length > 0)
                 .map((place, index) => (
@@ -35,90 +33,7 @@ export const NearbyPlaceList = ({ places, onSelectPlace }: Props) => {
                         onClick={() => onSelectPlace?.(place)}
                     />
                 ))}
-        </Container>
-    );
-};
-
-const Container = ({ children }: { children: ReactNode }) => {
-    const containerRef = useRef<HTMLDivElement>(null);
-
-    const scroll = (direction: "left" | "right") => {
-        if (!containerRef.current) {
-            return;
-        }
-        const container = containerRef.current;
-        const scrollLeft = container.scrollLeft;
-        const scrollAmount = 400;
-        const newScrollLeft =
-            scrollLeft + (direction === "left" ? -scrollAmount : scrollAmount);
-        container.scrollTo({ left: newScrollLeft, behavior: "smooth" });
-    };
-
-    return (
-        <Box position="relative" w="100%">
-            <HStack
-                ref={containerRef}
-                w="100%"
-                px={isPC ? "16px" : 0}
-                overflowX="auto"
-                overflowY="hidden"
-                scrollSnapType="x mandatory"
-                sx={{
-                    "::-webkit-scrollbar": {
-                        display: "none",
-                    },
-                }}
-            >
-                {children}
-            </HStack>
-            <Box>
-                <PageButton
-                    left={0}
-                    right="auto"
-                    onClick={() => scroll("left")}
-                >
-                    <Icon as={MdArrowBackIos} />
-                </PageButton>
-                <PageButton
-                    left="auto"
-                    right={0}
-                    onClick={() => scroll("right")}
-                >
-                    <Icon as={MdArrowForwardIos} />
-                </PageButton>
-            </Box>
-        </Box>
-    );
-};
-
-const PageButton = ({
-    left,
-    right,
-    onClick,
-    children,
-}: {
-    left: number | string;
-    right: number | string;
-    onClick?: () => void;
-    children: ReactNode;
-}) => {
-    return (
-        <Center
-            visibility={isPC ? "visible" : "hidden"}
-            position="absolute"
-            top="50%"
-            left={left}
-            right={right}
-            transform="translateY(-50%)"
-            w="40px"
-            h="40px"
-            backgroundColor="white"
-            borderRadius="100%"
-            boxShadow="0px 4px 8px rgba(0, 0, 0, 0.10)"
-            onClick={onClick}
-        >
-            {children}
-        </Center>
+        </HorizontaScrollablelList>
     );
 };
